requirements.txt
# Minimum requirements without heavy ML libraries
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0
redis==5.0.1
celery==5.3.6
openai==1.10.0
groq==0.4.1
google-generativeai==0.3.2
scikit-learn==1.4.0
sentence-transformers==2.3.1
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
duckduckgo-search==4.1.1
google-search-results==2.4.2
nltk==3.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
jinja2==3.1.3
loguru==0.7.2